+++ CHANGES.md
@@ -11,6 +11,7 @@ h2. 1.7.7
 
 * Parsing an empty hash with a representer having a wrap does no longer throw an exception.
 * `::nested` now works in modules, too! Nests are implemented as decorator representer, not as modules, so they don't pollute the represented object.
+* Introduce `:inherit` to allow inheriting+overriding properties and inline representers (and properties in inline representers - it starts getting crazy!!!).
 
 h2. 1.7.6